+++ granulate_utils/glogger/messages_buffer.py
@@ -1,3 +1,7 @@
+#
+# Copyright (c) Granulate. All rights reserved.
+# Licensed under the AGPL3 License. See LICENSE.md in the project root for license information.
+#
 import logging
 from typing import List
